+++ src/shared/index.ts
@@ -4,3 +4,5 @@
 
 export * from "./errors.js";
 export * from "./response.js";
+export * from "./tempfile.js";
+export * from "./merge.js";